% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBrCRM_index.R
\name{IBrCRM_index}
\alias{IBrCRM_index}
\title{Índice Brasileiro de Competitividade Regional Municipal (IBrCRM)}
\usage{
IBrCRM_index(
  df,
  variables,
  reference_variables,
  inverse_variables,
  group_by = NULL,
  adjust_outliers = TRUE,
  include_weight = TRUE,
  standardization_method = c("mean", "discrete", "none")
)
}
\arguments{
\item{variables}{A character vector specifying the axis style. Valid options are
\code{"none"} (no axis lines), \code{"full"} (full-length axis lines), and
\code{"half"} (half-length axis lines), the default.}

\item{reference_variables}{Logical value indicating whether to show text elements. If \code{TRUE},
axis text will be displayed in black; otherwise, they will
be hidden.}

\item{inverse_variables}{A character vector specifying the position of the
legend. Valid options are \code{"right"} (default), \code{"left"}, \code{"top"}, and
\code{"bottom"}.}

\item{group_by}{Defines whether the grid lines should be \code{"horizontal"}
(default) or \code{"vertical"}.}

\item{adjust_outliers}{Numeric. The number of breaks on the x-axis}

\item{include_weight}{Numeric. The number of breaks on the y-axis}

\item{standardization_method}{Logical value that indicates whether the x-axis
boundary should be expanded. If \code{TRUE}, the x-axis limits will be
expanded; otherwise there will be no change}
}
\value{
A \code{data.frame} output.
}
\description{
Creation of an inequality index capable of selecting important variables,
weights for each variable and normalizing the selected variables.
}
\examples{
library(dplyr)
variables <- c('mpg','cyl','hp','drat','wt','qsec','vs','am','gear')
reference_variables <- c('cyl','mpg')
group_by <- c('carb')
inverse_variables <- c('wt')

IBrCRM <- IBrCRM_index(mtcars,variables = variables, reference_variables = reference_variables,
                       inverse_variables = inverse_variables,
                       group_by = group_by,
                       adjust_outliers =TRUE, include_weight = TRUE,
                       standardization_method = c('mean'))
}
\concept{IBrCRM_index output}
