\name{plot_boruta_results}
\alias{plot_boruta_results}
\title{Visualiza o resultado do Boruta}
\usage{
plot_boruta_results(boruta_result)
}
\arguments{
\item{boruta_result}{Objeto retornado por \code{Boruta::Boruta()}.}
}
\value{
O valor retornado pelo método de gráfico do objeto Boruta, de forma invisível.
}
\description{
Produz um gráfico das importâncias estimadas pelo Boruta.
}
